+++ src/pyq.c
@@ -1,5 +1,5 @@
 #ifdef __linux__
-#define _GNU_SOURCE
+#  define _GNU_SOURCE
 #endif
 
 #include <unistd.h>
@@ -10,19 +10,75 @@
 #include <sys/types.h>
 #include <sys/stat.h>
 #include <pwd.h>
+#include <limits.h>
+#ifdef __APPLE__
+#  include <mach-o/dyld.h>
+#  include <uuid/uuid.h>
+#  include <sys/dirent.h>
+#endif /* __APPLE__ */
+
+#define TRACE if(pyq_trace) printf
+static int pyq_trace;
+static char progpath[PATH_MAX];
 
 #ifdef __APPLE__
-#include <mach-o/dyld.h>
-#include <uuid/uuid.h>
-#include <sys/dirent.h>
-
-#define QBIN "/m64/q"
-static char progpath[1024];
-static uint32_t nsexeclength = 1024;
-#endif
-#ifdef __linux__
-#define QBIN "/l64/q"
+static char*
+get_progpath(const char *progname)
+{
+    uint32_t nsexeclength = PATH_MAX;
+    /* TODO: Check for error and dynamically allocate progpath. */
+    _NSGetExecutablePath(progpath, &nsexeclength);
+    return strdup(progpath);
+}
+#elif defined(__linux__) /* __APPLE__ */
 #include <sched.h>
+static char *
+get_progpath(const char *progname)
+{
+    char* path;
+    char *b, *e, *r;
+    int n, m;
+    struct stat sb;
+    if (progname[0] == '/')
+        return strdup(progname);
+    if (progname[0] == '.' || strchr(progname, '/')) {
+        if (!getcwd(progpath, sizeof(progpath))) {
+            perror("getcwd");
+            exit(1);
+        }
+        if (!strncmp(progname, "./", 2)) {
+            strcat(progpath, progname + 1);
+            TRACE("Replace '.' in %s with cwd.  Got %s\n", progname, progpath);
+        }
+        else {
+            strcat(progpath, "/");
+            strcat(progpath, progname);
+            TRACE("Prepend %s with cwd.  Got %s\n", progname, progpath);
+        }
+        return strdup(progpath);
+    }
+    /* search from prog in the path */
+    path = getenv("PATH");
+    n = strlen(progname);
+    for (b = e = path; e; b = e + 1) {
+        e = strchr(b, ':');
+        if (e == NULL)
+            e = b + strlen(b);
+        m = e - b;
+        r = malloc(n + m + 2);
+        strncpy(r, b, m);
+        r[m] = '/';
+        strcpy(r + m + 1, progname);
+        if (stat(r, &sb) != -1) {
+            /* TODO: Check the executable bit. */
+            return r;
+        }
+        free(r);
+    }
+    /* If not found - leave as is */
+    return strdup(progname);
+}
+
 
 static cpu_set_t cpu_set;
 
@@ -34,29 +90,29 @@ parse_cpus(char *cpus)
     int i, j, cpu[2];
     CPU_ZERO(&cpu_set);
     for (str1 = cpus; ; str1 = NULL) {
-	token = strtok_r(str1, ",", &saveptr1);
-	if (token == NULL)
-	    break;
+        token = strtok_r(str1, ",", &saveptr1);
+        if (token == NULL)
+            break;
         for (i = 0, str2 = token; ; i++, str2 = NULL) {
-	    subtoken = strtok_r(str2, "-", &saveptr2);
-	    if (subtoken != NULL) {
-		cpu[i] = atoi(subtoken);
-		continue;
-	    }
-	    switch (i) {
-	    case 1:
-		CPU_SET(cpu[0], &cpu_set);
-		break;
-	    case 2:
-		for (j = cpu[0]; j <= cpu[1]; j++) {
-		    CPU_SET(j, &cpu_set);
-		}
-		break;
-	    default:
-		return -1;
-	    }
-	    break;
-	}
+            subtoken = strtok_r(str2, "-", &saveptr2);
+            if (subtoken != NULL) {
+                cpu[i] = atoi(subtoken);
+                continue;
+            }
+            switch (i) {
+            case 1:
+                CPU_SET(cpu[0], &cpu_set);
+                break;
+            case 2:
+                for (j = cpu[0]; j <= cpu[1]; j++) {
+                    CPU_SET(j, &cpu_set);
+                }
+                break;
+            default:
+                return -1;
+            }
+            break;
+        }
     }
     return 0;
 }
@@ -68,8 +124,8 @@ print_cpus(void)
     n = CPU_COUNT(&cpu_set);
     printf("n = %d\n", n);
     for (i = 0; i < CPU_SETSIZE; i++) {
-	if (CPU_ISSET(i, &cpu_set))
-	    printf("%d ", i);
+        if (CPU_ISSET(i, &cpu_set))
+            printf("%d ", i);
     }
     printf("\n");
 }
@@ -81,100 +137,123 @@ taskset(void)
     cpus = getenv("CPUS");
     test_cpus = getenv("TEST_CPUS");
     if (cpus) {
-	parse_cpus(cpus);
-	if (test_cpus) {
-	    print_cpus();
-	    exit(0);
-	}
-	sched_setaffinity(0, CPU_SETSIZE, &cpu_set);
+        parse_cpus(cpus);
+        if (test_cpus) {
+            print_cpus();
+            exit(0);
+        }
+        /* From man sched_setaffinity:
+
+        The argument cpusetsize is the length (in bytes) of
+        the data pointed to by mask. Normally this argument
+        would  be  specified  as sizeof(cpu_set_t). */
+        sched_setaffinity(0, sizeof(cpu_set_t), &cpu_set);
     }
     return 0;
 }
+#else /* __linux__ */
+    #error "Unsupported OS"
 #endif
 
-#ifndef QBIN
-#error "Cannot figure out QBIN"
+#ifndef QARCH
+#error "QARCH is not set."
 #endif
-#define QHBIN "/q" QBIN
 
-static int
-no_q(char *path)
-{
-    int result;
-    struct stat st;
-    char *q_path;
-    q_path = malloc(strlen(path) + 12);
-    strcpy(q_path, path);
-    strcat(q_path, "/q/python.q");
-    result = stat(q_path, &st) != 0;
-    /* Add more checks for a valid QHOME */
-    free(q_path);
-    return result;
-}
+#define xstr(s) str(s)
+#define str(s) #s
 
+/* QHOME priority:
+    1. QHOME is already set in the environment.
+    2. VIRTUAL_ENV is set -> QHOME/$VIRTUAL_ENV/q
+    3. Next to pyq executable
+    4. ~/q
+*/
 char *
-find_q(void)
+find_q(const char* progpath)
 {
-    char *qhome, *home, *qpath, *venv;
-    int pathlen;
-    qhome = getenv("QHOME");
-    venv = getenv("VIRTUAL_ENV");
-    if (qhome == NULL) {
-        if (venv && no_q(venv)) {
-            venv = NULL;
-        }
-        if (venv == NULL) {
-            home = getenv("HOME");
-            if (home == NULL) {
-                struct passwd *pw = getpwuid(getuid());
-                home = pw->pw_dir;
+    char *qhome = NULL, *prefix = NULL, *qpath = NULL, *p;
+    static int attempt = 0;
+    switch (++attempt) {
+        case 1:
+            if ((qhome = getenv("QHOME"))) {
+                qhome = strdup(qhome);
+                break;
             }
-            pathlen = strlen(home) + strlen(QHBIN) + 1;
-        }
-        else {
-            pathlen = strlen(venv) + strlen(QHBIN) + 1;
-        }
+            ++attempt;
+            /* fall through */
+        case 2:
+            if ((prefix = getenv("VIRTUAL_ENV"))) {
+                qhome = malloc(strlen(prefix) + 3);
+                strcpy(qhome, prefix);
+                strcat(qhome, "/q");
+                setenv("QHOME", strdup(qhome), 1);
+                break;
+            }
+            ++attempt;
+        case 3:
+            p = strrchr(progpath, '/');
+            if (p && (p -= 4) > progpath && !strncmp(p, "/bin/", 5)) {
+                int n = p - progpath;
+                qhome = malloc(n + 3);
+                strncpy(qhome, progpath, n);
+                strcat(qhome, "/q");
+                setenv("QHOME", strdup(qhome), 1);
+                break;
+            }
+            ++attempt;
+            /* fall through */
+        case 4:
+            prefix = getenv("HOME");
+            if (prefix == NULL) {
+                struct passwd *pw = getpwuid(getuid());
+                prefix = pw->pw_dir;
+            }
+            if (prefix) {
+                qhome = malloc(strlen(prefix) + 3);
+                strcpy(qhome, prefix);
+                strcat(qhome, "/q");
+                break;
+            }
+             ++attempt;
+             /* fall through */
+        default:
+            return NULL;
     }
-    else {
-        pathlen = strlen(qhome) + strlen(QBIN) + 1;
-    }
-    qpath = malloc(pathlen);
-    if (qhome == NULL) {
-        if (venv) {
-            strcpy(qpath, venv);
-            strcat(qpath, "/q");
-            setenv("QHOME", strdup(qpath), 1);
-            strcat(qpath, QBIN);
-        }
-        else {
-            strcpy(qpath, home);
-            strcat(qpath, QHBIN);
-        }
-    }
-    else {
-        strcpy(qpath, qhome);
-        strcat(qpath, QBIN);
-    }
+    qpath = malloc(strlen(qhome) + strlen(xstr(QARCH)) + 4);
+    strcpy(qpath, qhome);
+    strcat(qpath, "/" xstr(QARCH) "/q");
+    setenv("QBIN", qpath, 1);
 
     return qpath;
 }
 
+#define NPATHS 4
+
 int
 main(int argc, char *argv[])
 {
-    int rc, i;
+    char *tried_paths[NPATHS];
+    int rc = 0, i, n;
     char **args, *p, *qpath;
-    qpath = find_q();
-    setenv("QBIN", qpath, 1);
+    char *fullprogpath;
+    pyq_trace = argc > 1 && !strcmp("--pyq-trace", argv[1]);
+    if (pyq_trace) {
+        /* remove option from argv */
+        --argc;
+        for (i = 1; i < argc; ++i) {
+            argv[i] = argv[i+1];
+        }
+        argv[argc] = NULL;
+        printf("pyq trace is on\n");
+    }
     args = malloc((sizeof (char*)) * (argc + 3));
+    fullprogpath = get_progpath(argv[0]);
 #ifdef __APPLE__
-    /* TODO: Check for error and dynamically allocate progpath. */
-    _NSGetExecutablePath(progpath, &nsexeclength);
+
     args[0] = progpath;
     setenv("PYTHONEXECUTABLE", progpath, 1);
-#else
-    args[0] = argv[0];
 #endif
+    args[0] = fullprogpath;
     args[1] = "python.q";
     args[argc + 1] = "-q";
     args[argc + 2] = NULL;
@@ -189,17 +268,39 @@ main(int argc, char *argv[])
             args[i+1] = argv[i];
         }
     }
+    TRACE("prog = %s\n", fullprogpath);
 #ifdef __linux__
     taskset();
 #endif
-    rc = execvp(qpath, args);
-    /* we can only get here on error */
-    i = strlen(qpath) - strlen(QBIN);
-    strncpy(qpath + i + 2, "32", 2);
-    setenv("QBIN", qpath, 1);
-    rc = execvp(qpath, args);
-    /* we can only get here on error */
-    perror(qpath);
 
+    for(i = 0; ;++i) {
+        qpath = find_q(args[0]);
+        if (qpath) {
+            TRACE("qbin = %s\n", qpath);
+            if (pyq_trace) {
+                printf("args =");
+                char **p;
+                for (p = args; *p; ++p) {
+                    printf(" %s", *p);
+                }
+                printf("\n");
+
+            }
+            /* Flush streams before exec. */
+            fflush(stdout);
+            fflush(stderr);
+            rc = execvp(qpath, args);
+            tried_paths[i] = qpath;
+        }
+        else {
+            break;
+        }
+    }
+    /* we can only get here on error */
+    if (!pyq_trace) /* In verbose mode, paths are already printed */
+        for (n = i, i = 0; i < n; ++i) {
+            fprintf(stderr, "qbinpath = %s\n", tried_paths[i]);
+        }
+    perror(qpath);
     return rc;
 }